+++ quickbooks/objects/detailline.py
@@ -13,7 +13,7 @@ class DetailLine(QuickbooksBaseObject):
         super(DetailLine, self).__init__()
         self.Id = None
         self.LineNum = 0
-        self.Description = ""
+        self.Description = None
         self.Amount = 0
         self.DetailType = ""
         self.LinkedTxn = []
@@ -135,12 +135,121 @@ class SalesItemLineDetail(QuickbooksBaseObject):
         return str(self.UnitPrice)
 
 
-class SaleItemLine(DetailLine):
+class SalesItemLine(DetailLine):
     class_dict = {
         "SalesItemLineDetail": SalesItemLineDetail
     }
 
     def __init__(self):
-        super(SaleItemLine, self).__init__()
+        super(SalesItemLine, self).__init__()
         self.DetailType = "SalesItemLineDetail"
         self.SalesItemLineDetail = None
+
+
+class GroupLineDetail(QuickbooksBaseObject):
+    pass
+
+
+class GroupLine(DetailLine):
+    class_dict = {
+        "GroupLineDetail": GroupLineDetail
+    }
+
+    def __init__(self):
+        super(GroupLine, self).__init__()
+        self.DetailType = "SalesItemLineDetail"
+        self.SalesItemLineDetail = None
+
+
+class DescriptionLineDetail(QuickbooksBaseObject):
+    class_dict = {
+        "TaxCodeRef": Ref,
+    }
+
+    def __init__(self):
+        super(DescriptionLineDetail, self).__init__()
+        self.ServiceDate = ""
+        self.TaxCodeRef = None
+
+
+class DescriptionOnlyLine(DetailLine):
+    class_dict = {
+        "DescriptionLineDetail": DescriptionLineDetail
+    }
+
+    def __init__(self):
+        super(DescriptionOnlyLine, self).__init__()
+        self.DetailType = "DescriptionLineDetail"
+        self.DescriptionLineDetail = None
+
+
+@python_2_unicode_compatible
+class AccountBasedExpenseLineDetail(QuickbooksBaseObject):
+    class_dict = {
+        "CustomerRef": Ref,
+        "AccountRef": Ref,
+        "TaxCodeRef": Ref,
+        "ClassRef": Ref,
+        "MarkupInfo": MarkupInfo,
+    }
+
+    def __init__(self):
+        super(AccountBasedExpenseLineDetail, self).__init__()
+        self.BillableStatus = None
+        self.TaxAmount = 0
+        self.TaxInclusiveAmt = 0
+
+        self.CustomerRef = None
+        self.AccountRef = None
+        self.TaxCodeRef = None
+
+    def __str__(self):
+        return self.BillableStatus
+
+
+class AccountBasedExpenseLine(DetailLine):
+    class_dict = {
+        "AccountBasedExpenseLineDetail": AccountBasedExpenseLineDetail
+    }
+
+    def __init__(self):
+        super(AccountBasedExpenseLine, self).__init__()
+
+        self.DetailType = "AccountBasedExpenseLineDetail"
+        self.AccountBasedExpenseLineDetail = None
+
+
+class ItemBasedExpenseLineDetail(QuickbooksBaseObject):
+    class_dict = {
+        "ItemRef": Ref,
+        "ClassRef": Ref,
+        "PriceLevelRef": Ref,
+        "TaxCodeRef": Ref,
+        "CustomerRef": Ref,
+        "MarkupInfo": MarkupInfo
+    }
+
+    def __init__(self):
+        super(ItemBasedExpenseLineDetail, self).__init__()
+        self.BillableStatus = ""
+        self.UnitPrice = 0
+        self.TaxInclusiveAmt = 0
+        self.Qty = 0
+        self.ItemRef = None
+        self.ClassRef = None
+        self.PriceLevelRef = None
+        self.TaxCodeRef = None
+        self.MarkupInfo = None
+        self.CustomerRef = None
+
+
+class ItemBasedExpenseLine(DetailLine):
+    class_dict = {
+        "ItemBasedExpenseLineDetail": ItemBasedExpenseLineDetail
+    }
+
+    def __init__(self):
+        super(ItemBasedExpenseLine, self).__init__()
+
+        self.DetailType = "ItemBasedExpenseLineDetail"
+        self.ItemBasedExpenseLineDetail = None