+++ quickbooks/objects/refundreceipt.py
@@ -29,6 +29,10 @@ class RefundReceipt(QuickbooksManagedObject, QuickbooksTransactionEntity, Linked
         "LinkedTxn": LinkedTxn
     }
 
+    detail_dict = {
+
+    }
+
     qbo_object_name = "RefundReceipt"
 
     def __init__(self):