+++ tests/unit/objects/test_bill.py
@@ -1,15 +1,8 @@
 import unittest
 
 from quickbooks import QuickBooks
-from quickbooks.objects.bill import Bill, BillLine, AccountBasedExpenseLineDetail, ItemBasedExpenseLineDetail
+from quickbooks.objects.bill import Bill
 
-
-class AccountBasedExpenseLineDetailTests(unittest.TestCase):
-    def test_unicode(self):
-        acct_detail = AccountBasedExpenseLineDetail()
-        acct_detail.BillableStatus = "test"
-
-        self.assertEquals(str(acct_detail), "test")
 
 
 class BillTests(unittest.TestCase):
@@ -47,26 +40,3 @@ class BillTests(unittest.TestCase):
         self.assertEquals(ref.type, "Bill")
         self.assertEquals(ref.value, 100)
 
-
-class BillLineTests(unittest.TestCase):
-    def test_unicode(self):
-        bill_line = BillLine()
-        bill_line.Amount = 1000
-
-        self.assertEquals(str(bill_line), "1000")
-
-
-class ItemBasedExpenseLineDetailTest(unittest.TestCase):
-    def test_init(self):
-        detail = ItemBasedExpenseLineDetail()
-
-        self.assertEquals(detail.BillableStatus, "")
-        self.assertEquals(detail.UnitPrice, 0)
-        self.assertEquals(detail.TaxInclusiveAmt, 0)
-        self.assertEquals(detail.Qty, 0)
-        self.assertEquals(detail.ItemRef, None)
-        self.assertEquals(detail.ClassRef, None)
-        self.assertEquals(detail.PriceLevelRef, None)
-        self.assertEquals(detail.TaxCodeRef, None)
-        self.assertEquals(detail.MarkupInfo, None)
-        self.assertEquals(detail.CustomerRef, None)