+++ tests/unit/objects/test_creditmemo.py
@@ -1,7 +1,8 @@
 import unittest
 
 from quickbooks import QuickBooks
-from quickbooks.objects.creditmemo import SalesItemLineDetail, CreditMemoLine, CreditMemo, \
+from quickbooks.objects.creditmemo import CreditMemo
+from quickbooks.objects.detailline import SalesItemLineDetail, \
     DiscountLineDetail, SubtotalLineDetail, DiscountOverride, DescriptionLineDetail
 
 
@@ -11,16 +12,6 @@ class SalesItemLineDetailTests(unittest.TestCase):
         detail.UnitPrice = 10
 
         self.assertEquals(str(detail), "10")
-
-
-class CreditMemoLineTests(unittest.TestCase):
-    def test_unicode(self):
-        memo_line = CreditMemoLine()
-        memo_line.LineNum = 1
-        memo_line.Description = "Product Description"
-        memo_line.Amount = 100
-
-        self.assertEquals(str(memo_line), "[1] Product Description 100")
 
 
 class CreditMemoTests(unittest.TestCase):