+++ tests/unit/test_mixins.py
@@ -3,8 +3,9 @@ from mock import patch, Mock
 
 from quickbooks import client
 
-from quickbooks.objects.base import PhoneNumber
+from quickbooks.objects.base import PhoneNumber, QuickbooksBaseObject
 from quickbooks.objects.department import Department
+from quickbooks.objects.journalentry import JournalEntry, JournalEntryLine
 from quickbooks.objects.salesreceipt import SalesReceipt
 
 
@@ -16,6 +17,49 @@ class ToJsonMixinTest(unittest.TestCase):
         json = phone.to_json()
 
         self.assertEquals(json, '{\n    "FreeFormNumber": "555-555-5555"\n}')
+
+
+class FromJsonMixinTest(unittest.TestCase):
+    def setUp(self):
+        self.json_data = {
+            'DocNumber': '123',
+            'TotalAmt': 100,
+            'Line': [
+                {
+                    "Id": "0",
+                    "Description": "Test",
+                    "Amount": 25.54,
+                    "DetailType": "JournalEntryLineDetail",
+                    "JournalEntryLineDetail": {
+                        "PostingType": "Debit",
+                    }
+                },
+            ],
+        }
+
+    def test_from_json(self):
+        entry = JournalEntry()
+        new_obj = entry.from_json(self.json_data)
+
+        self.assertEquals(type(new_obj), JournalEntry)
+        self.assertEquals(new_obj.DocNumber, "123")
+        self.assertEquals(new_obj.TotalAmt, 100)
+
+        line = new_obj.Line[0]
+        self.assertEquals(type(line), JournalEntryLine)
+        self.assertEquals(line.Description, "Test")
+        self.assertEquals(line.Amount, 25.54)
+        self.assertEquals(line.DetailType, "JournalEntryLineDetail")
+        self.assertEquals(line.JournalEntryLineDetail.PostingType, "Debit")
+
+    def test_from_json_missing_detail_object(self):
+        test_obj = QuickbooksBaseObject()
+
+        new_obj = test_obj.from_json(self.json_data)
+
+        self.assertEquals(type(new_obj), QuickbooksBaseObject)
+        self.assertEquals(new_obj.DocNumber, "123")
+        self.assertEquals(new_obj.TotalAmt, 100)
 
 
 class ListMixinTest(unittest.TestCase):
@@ -53,7 +97,8 @@ class ListMixinTest(unittest.TestCase):
     @patch('quickbooks.mixins.ListMixin.query')
     def test_where(self, query):
         Department.where("Active=True", 1, 10)
-        query.assert_called_once_with("SELECT * FROM Department WHERE Active=True STARTPOSITION 1 MAXRESULTS 10", qb=None)
+        query.assert_called_once_with("SELECT * FROM Department WHERE Active=True STARTPOSITION 1 MAXRESULTS 10",
+                                      qb=None)
 
     def test_where_with_qb(self):
         with patch.object(self.qb_client, 'query') as query:
@@ -121,13 +166,13 @@ class UpdateMixinTest(unittest.TestCase):
     @patch('quickbooks.mixins.QuickBooks.create_object')
     def test_save_create(self, create_object):
         department = Department()
-        department.save()
+        department.save(qb=self.qb_client)
         create_object.assert_called_once_with("Department", department.to_json())
 
     def test_save_create_with_qb(self):
         with patch.object(self.qb_client, 'create_object') as create_object:
             department = Department()
-            department.save()
+            department.save(qb=self.qb_client)
             self.assertTrue(create_object.called)
 
     @patch('quickbooks.mixins.QuickBooks.update_object')
@@ -136,7 +181,7 @@ class UpdateMixinTest(unittest.TestCase):
         department.Id = 1
         json = department.to_json()
 
-        department.save()
+        department.save(qb=self.qb_client)
         update_object.assert_called_once_with("Department", json)
 
     def test_save_update_with_qb(self):
@@ -145,7 +190,7 @@ class UpdateMixinTest(unittest.TestCase):
             department.Id = 1
             json = department.to_json()
 
-            department.save()
+            department.save(qb=self.qb_client)
             self.assertTrue(update_object.called)