+++ CHANGELOG.rst
@@ -1,8 +1,7 @@
 Changelog
 ========
 
-* 0.5.0 (July 22, 2016)
-    * Fixed bug on PurchaseLine.
+* 0.5.0 (July 25, 2016)
     * Added ability to query current user.
     * Added support to reconnect an account.
     * Added to_ref method to Bill object.
@@ -17,6 +16,12 @@ Changelog
     * Fixed issues that prevented save from working on TaxService.
     * Removed unsupported save method from TaxRate.
     * Removed unsupported save method from TaxCode.
+    * Added to_ref method to TaxCode.
+    * Fixed issues with default values on Estimate.
+    * Fixed issues loading detail lines on the following objects: JournalEntry, CreditMemo, Bill, Purchase and PurchaseOrder.
+    * Corrected spelling of object SaleItemLine to SalesItemLine.
+    * Removed the following objects: CreditMemoLine, BillLine, JournalEntryLine, PurchaseLine, and PurchaseOrderLine.
+
 
 * 0.4.0 (June 15, 2016)
     * Added a way of disconnecting a Quickbooks Account to client.